Collatz Conjecture Calculator Upload.
The Collatz Conjecture, is one of the seemingly simplest, yet most complex and fascinating mathematical problems I've took an intellectual interest in learning.

I've decided to embark on the journey of simplifying the calculation of the numbers' sequences until they reach the [4,2,1] loop, by printing said sequence, once the user has inputted the proper positive integer necessary to perform these calculations.

Essentially, the user is instructed to input a positive integer, once the program has been launched. Then, by pressing enter, the program determines whether the number is even or odd. If it is even, it uses the function 3X+1 to print out the new number Y. If it's odd, then it uses the X/2 function instead. Then, it runs a check on the new obtained number is even or odd, and follows the same procedure until the number gets stuck in the 4,2,1 loop. And this is when the program stops.

For further clarification, I highly recommend viewing an explanation video on what the Collatz Conjecture actually is; in order to get a solid understanding on why such a program is fundamentally necessary for the efficiency of effort being put by new learners into studying this mathematical beauty.