Age Calculator Upload
Uploaded the C++ Code for another project of mine. The Age Calculator delivers accurate and spot-on results, after the user inputs their age, and current year. The code runs basic functions to determine what said person's year of birth is, in addition to how many months, or days it has been since said person was born. Further information could be deduced through the code itself, and the comments listed in the source code.