Bytes Converter Upload
Uploaded a C++ Program used to convert the number of bytes into other units of measurement. Said converter asks the user for the number of kilobytes that they'd like to convert. Upon inserting the number, the code uses spot-on conversion formulas to deliver the number of bytes, and bits in said amount of kilobytes. The precision, clarity, and user-friendly theme of the program makes it easier for this task to be fulfilled.